feat: global search via unified /api/search + cache seed tooling (R3)
Frontend refactor (behavior-neutral for the search pages):
- The pure SPARQL template functions move from QueryService to a new
  service/query.templates.js with no Nuxt/Pinia imports, so Node tooling
  can generate byte-identical queries to the ones the app sends (the
  backend cache is keyed by the exact query text). QueryService delegates.
- The 8 search-form definitions move from pages/search/*/query.vue to
  plain factories in service/search-forms/<table>.js, importable from
  both the pages and the seed generator.

Global search: Simple.vue now posts a globalSearchQuery (all 8 tables,
aliases GROUP_CONCATed per item+pbid, searchVars label,aliases,pbid,item
so pbid/alias/Q-number matches keep working) to /api/search v=2 via the
new WikibaseService.cachedSearch, replacing quickSearch. The backend
/api/search/quick endpoint stays untouched as an alias for cached SPAs
until R5.

Seed tooling (scripts/seed-cache/): generate-queries.mjs imports the
real frontend templates and form definitions, fetches the SPARQL prefix
from the target backend's /api/config, dedupes by final query text and
emits queries.json (default scope deliberately conservative: 5 global
queries + group=ALL autocompletes, 291 entries); seed.sh fires them and
re-polls indexLoading until every index is materialized.

Verified end-to-end: seeded a fresh backend (global.en = 155k rows in
~4 min), then drove the real UI — geoid autocompletes and the global
search bar hit the seeded entries with zero hash fragmentation
(registry count unchanged), pbid search works, selecting a result
navigates to its item page.
